Refactor Lib Classes and add minor changes
