+++ src/Lib/Bdd.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Model;
+namespace App\Lib;
 
 use PDO;
