+++ src/Lib/Pagination.php
@@ -6,13 +6,15 @@ class Pagination
 {
     private $total;
     private $perPage;
+    private $currentPage;
 
-    public function __construct($total, int $perPage = 10)
+    public function __construct($total, $currentPage)
 
     {
 
         $this->total = $total;
-        $this->perPage = $perPage;
+        $this->perPage = 50;
+        $this->currentPage = $currentPage;
     }
 
     public function getPerPage()
@@ -24,9 +26,6 @@ class Pagination
     public function getCurrentPage()
     {
 
-        if (isset($_GET['id']) && !empty($_GET['id'])) {
-            $this->currentPage = (int) strip_tags($_GET['id']);
-        }
         return $this->currentPage;
     }
