+++ src/Lib/Route.php
@@ -35,7 +35,6 @@ class Route
         $params = explode('@', $this->action);
         $controller = new $params[0]();
         $method = $params[1];
-
         return isset($this->matches[1]) ? $controller->$method($this->matches[1]) : $controller->$method();
     }
 }
